доработка заданий 5 и 6
